+++ pypots/imputation/autoformer/modules/core.py
@@ -7,9 +7,7 @@
 
 import torch.nn as nn
 
-from ....utils.metrics import calc_mse
 from .submodules import (
-    DataEmbedding_wo_Pos,
     SeriesDecompositionBlock,
     SeasonalLayerNorm,
     AutoformerEncoderLayer,
@@ -17,6 +15,8 @@ from .submodules import (
     AutoCorrelation,
     AutoCorrelationLayer,
 )
+from ....nn.modules.transformer.embedding import DataEmbedding
+from ....utils.metrics import calc_mse
 
 
 class _Autoformer(nn.Module):
@@ -39,10 +39,11 @@ class _Autoformer(nn.Module):
         self.seq_len = n_steps
         self.n_layers = n_layers
         self.series_decomp = SeriesDecompositionBlock(moving_avg_window_size)
-        self.enc_embedding = DataEmbedding_wo_Pos(
+        self.enc_embedding = DataEmbedding(
             n_features,
             d_model,
             dropout=dropout,
+            with_pos=False,
         )
         self.encoder = AutoformerEncoder(
             [
