+++ pypots/nn/modules/transformer/embedding.py
@@ -1,9 +1,13 @@
 """
+Embedding methods for Transformer models are put here.
 
+
+Partial implementation uses code from https://github.com/zhouhaoyi/Informer2020/blob/main/models/embed.py
 """
 
 # Created by Wenjie Du <wenjay.du@gmail.com>
 # License: BSD-3-Clause
+
 
 import math
 
@@ -11,7 +15,62 @@ import torch
 import torch.fft
 import torch.nn as nn
 
-from ....nn.modules.transformer import PositionalEncoding
+
+class PositionalEncoding(nn.Module):
+    """The original positional-encoding module for Transformer.
+
+    Parameters
+    ----------
+    d_hid:
+        The dimension of the hidden layer.
+
+    n_positions:
+        The max number of positions.
+
+    """
+
+    def __init__(self, d_hid: int, n_positions: int = 1000):
+        super().__init__()
+        pe = torch.zeros(n_positions, d_hid, requires_grad=False).float()
+        position = torch.arange(0, n_positions).float().unsqueeze(1)
+        div_term = (
+            torch.arange(0, d_hid, 2).float()
+            * -(torch.log(torch.tensor(10000)) / d_hid)
+        ).exp()
+
+        pe[:, 0::2] = torch.sin(position * div_term)
+        pe[:, 1::2] = torch.cos(position * div_term)
+
+        pe = pe.unsqueeze(0)
+        self.register_buffer("pos_table", pe)
+
+    def forward(self, x: torch.Tensor, return_only_pos: bool = False) -> torch.Tensor:
+        """Forward processing of the positional encoding module.
+
+        Parameters
+        ----------
+        x:
+            Input tensor.
+
+        return_only_pos:
+            Whether to return only the positional encoding.
+
+        Returns
+        -------
+        If return_only_pos is True:
+            pos_enc:
+                The positional encoding.
+        else:
+            x_with_pos:
+                Output tensor, the input tensor with the positional encoding added.
+        """
+        pos_enc = self.pos_table[:, : x.size(1)].clone().detach()
+
+        if return_only_pos:
+            return pos_enc
+
+        x_with_pos = x + pos_enc
+        return x_with_pos
 
 
 class TokenEmbedding(nn.Module):
@@ -103,11 +162,22 @@ class TimeFeatureEmbedding(nn.Module):
 
 
 class DataEmbedding(nn.Module):
-    def __init__(self, c_in, d_model, embed_type="fixed", freq="h", dropout=0.1):
+    def __init__(
+        self,
+        c_in,
+        d_model,
+        embed_type="fixed",
+        freq="h",
+        dropout=0.1,
+        with_pos=True,
+    ):
         super().__init__()
 
+        self.with_pos = with_pos
+
         self.value_embedding = TokenEmbedding(c_in=c_in, d_model=d_model)
-        self.position_embedding = PositionalEncoding(d_hid=d_model)
+        if with_pos:
+            self.position_embedding = PositionalEncoding(d_hid=d_model)
         self.temporal_embedding = (
             TemporalEmbedding(d_model=d_model, embed_type=embed_type, freq=freq)
             if embed_type != "timeF"
@@ -117,13 +187,11 @@ class DataEmbedding(nn.Module):
 
     def forward(self, x, x_timestamp=None):
         if x_timestamp is None:
-            x = self.value_embedding(x) + self.position_embedding(
-                x, return_only_pos=True
-            )
+            x = self.value_embedding(x)
+            if self.with_pos:
+                x += self.position_embedding(x, return_only_pos=True)
         else:
-            x = (
-                self.value_embedding(x)
-                + self.temporal_embedding(x_timestamp)
-                + self.position_embedding(x, return_only_pos=True)
-            )
+            x = self.value_embedding(x) + self.temporal_embedding(x_timestamp)
+            if self.with_pos:
+                x += self.position_embedding(x, return_only_pos=True)
         return self.dropout(x)
